Liam

+++ src/main/java/se/aljr/application/Launcher.java
@@ -23,7 +23,7 @@ public class Launcher {
             if(!isLoggedIn){
 
                 loginWindow.dispose();
-                ApplicationWindow applicationWindow = new ApplicationWindow(Monitorsize.getWidth()/2, Monitorsize.getHeight()/2, "Brogress - Gym tracker");
+                ApplicationWindow applicationWindow = new ApplicationWindow((int)(Monitorsize.getWidth()/2), Monitorsize.getHeight()/2, "Brogress - Gym tracker");
                 break;
             }
         }
